gui static
decided to go with a static file for the gui instead since I can listen and whatnot on just that file. This way the gui and server are completely separate. Also made the server available to the gui routine so it can stop and restart the server as necessiary if they update the path or port.